re-org and clean up messaging app
