+++ examples/messaging/src/assets/index.js
@@ -1,0 +1,4 @@
+export { CommandIcon } from './CommandIcon';
+export { CreateChannelIcon } from './CreateChannelIcon';
+export { EmojiIcon } from './EmojiIcon';
+export { SendIcon } from './SendIcon';
