--- examples/messaging/src/components/MessagingChannelHeader.css
@@ -1,33 +1,0 @@
-
-.messaging__channel-header {
-    background: rgba(255, 255, 255, 0.9);
-    box-shadow: 0px 7px 9px rgba(0, 0, 0, 0.03), 0px 1px 0px rgba(0, 0, 0, 0.03);
-    transform: matrix(1, 0, 0, 1, 0, 0);
-    border-radius: 10px 10px 0px 0px;
-    display: flex;
-    align-items: center;
-    min-height: 60px;
-    padding: 16px;
-  }
-
-  .messaging__channel-header .channel-header__name {
-    height: 15px;
-    flex: 1;
-    font-style: normal;
-    font-weight: bold;
-    font-size: 15px;
-    line-height: 100%;
-    color: rgba(0, 0, 0, 0.9);
-    text-transform: capitalize;
-  }
-
-
-.str-chat.dark .messaging__channel-header {
-  background: rgba(46, 48, 51, 0.98);
-  box-shadow: 0px 7px 9px rgba(0, 0, 0, 0.03), 0px 1px 0px rgba(0, 0, 0, 0.03);
-}
-
-.str-chat.dark .messaging__channel-header .channel-header__name {
-  color: rgba(255, 255, 255, 0.9);
-}
-
