--- examples/messaging/src/components/MessagingCreateChannel.css
@@ -1,132 +1,0 @@
-.messaging-create-channel {
-    width: 100%;
-    height: 100%;
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    flex: 1 1;
-    padding: 20px 20px 0 0px;
-}
-
-
-
-.messaging-create-channel header {
-    background: rgba(255, 255, 255, 0.9);
-    box-shadow: 0px 7px 9px rgba(0, 0, 0, 0.03), 0px 1px 0px rgba(0, 0, 0, 0.03);
-    transform: matrix(1, 0, 0, 1, 0, 0);
-    border-radius: 10px 10px 0px 0px;
-    display: flex;
-    align-items: center;
-    height: 64px;
-    padding: 0 16px;
-}
-
-.messaging-create-channel main {
-    flex: 1;
-    background-color: white;
-}
-
-
-
-
-.messaging-create-channel__users {
-    display: flex;
-    margin:  0 8px;
-}
-.messaging-create-channel__user {
-    color: white;
-    background: #005fff;
-    display: block;
-    padding: 6px 24px 6px 12px;
-    border-radius: 20px;
-    cursor: pointer;
-}
-.messaging-create-channel__user:not(:last-of-type) {
-    margin-right: 4px;
-}
-
-.messaging-create-channel__input {
-    height: 100%;
-    background: none;
-    border: none;
-}
-
-.messaging-create-channel__user-results {    
-    width: fit-content;
-    min-width: 375px;
-    background: #282A2D;
-    box-shadow: 0px 4px 20px rgba(0, 0, 0, 0.3);    border-radius: 10px;
-    overflow-y: auto;
-    max-height: 100%;
-    list-style: none;
-    padding: 0;
-    margin-left: 16px;
-}
-
-.messaging-create-channel__user-result {
-    height: 56px;
-    display: flex;
-    align-items: center;
-    margin: 0;
-    padding: 8px 16px;
-    cursor: pointer;
-}
-
-.messaging-create-channel__user-result__avatar {
-    
-}
-
-.messaging-create-channel__user-result__details {
-    display: flex;
-    flex-direction: column;
-}
-
-.messaging-create-channel__user-result__details__last-seen {
-    font-weight: normal;
-    font-size: 13px;
-    line-height: 16px;
-}
-
-/* modifiers dark theme */
-.str-chat.dark .messaging-create-channel__input {
-    height: 100%;
-    background: none;
-    border: none;
-    color: rgba(255,255,255,0.9);
-    outline: none;
-}
-
-.str-chat.dark .messaging-create-channel header {
-    background: rgba(46, 48, 51, 0.98);
-    box-shadow: 0px 7px 9px rgba(0, 0, 0, 0.03), 0px 1px 0px rgba(0, 0, 0, 0.03);
-    color: rgba(255, 255, 255, 0.9);
-
-}
-.str-chat.dark .messaging-create-channel main {
-    background: rgba(255, 255, 255, 0.04);
-}
-
-.str-chat.dark .messaging-create-channel__user-result:hover, .str-chat.dark .messaging-create-channel__user-result.selected {
-    background-color: #3E3E41;
-}
-
-.str-chat.dark .messaging-create-channel__user-result__details {
-    color: white;
-}
-
-.str-chat.dark .messaging-create-channel__user-result:hover, .str-chat.dark .messaging-create-channel__user-result.selected {
-    background-color: #3E3E41;
-}
-
-.str-chat.dark .messaging-create-channel__user-result__details__last-seen {
-    color: #858688;
-}
-
-/* modifiers page structure */
-.str-chat-channel.messaging .messaging-create-channel + .str-chat__main-panel {
-    display: none;
-}
-
-.str-chat-channel.messaging .messaging-create-channel ~ .str-chat__thread {
-    display: none;
-}
